backend/requirements-minimal.txt
# Minimal requirements for emergency deployment
fastapi==0.116.1
uvicorn[standard]==0.35.0
python-dotenv==1.1.1
pymongo[srv]==4.9.2
google-generativeai==0.8.5
pinecone==7.3.0
gunicorn==21.2.0
requests>=2.32.3
numpy>=1.24.0
